Just the visualization left
I've formatted the notebook, added the introduction, methods, expected outcomes and significance from the google docs. Beside, I've added some notes between steps to help solidify the understanding of the process. I've also split the data and used the training data to summarize the observations for each "result", find the number of NA values for each column and the average value for each variable. The visualization part is what remained to be done.